Refactored help, fixed bug on point outside of CDF, now support multiple days files
