+++ src/model/model-process.ts
@@ -194,107 +194,121 @@ export const processCompetition = (competition: Competition, datas: any[][]) => 
 
   datas
     .filter((data) => data)
-    .forEach((data: any[]) => {
-      // add new day
-      const day = Number(data[0].Jo);
-      const dayCompetition: CompetitionDay = {
-        day,
-        teams: [],
-        matchs: [],
-        pools: new Map(),
-      };
-      competition.dayCount = day;
-      competition.days[day] = dayCompetition;
-      // lastDays are updated when adding a played match
+    .map((data: any[]) => {
+      // split multiple days in different arrays to support multiple days file
+      const split: any[][] = [];
+      data.forEach((match: any) => {
+        const day = Number(match.Jo);
+        if (!split[day]) {
+          split[day] = data.filter((dayMatch: any) => dayMatch.Jo === match.Jo);
+        }
+      });
+      return split;
+      // return [data];
+    })
+    .forEach((daydata: any[][]) => {
+      daydata.forEach((data: any[]) => {
+        // add new day
+        const day = Number(data[0].Jo);
+        const dayCompetition: CompetitionDay = {
+          day,
+          teams: [],
+          matchs: [],
+          pools: new Map(),
+        };
+        competition.dayCount = day;
+        competition.days[day] = dayCompetition;
+        // lastDays are updated when adding a played match
 
-      // process day
-      if (!isCDF) {
-        data.forEach((match: any) => {
-          const teamA = getTeam(competition, match.EQA_no, match.EQA_nom);
-          const teamB = getTeam(competition, match.EQB_no, match.EQB_nom);
-          [teamA, teamB].forEach((team: Team) => {
-            dayCompetition.teams.push(team);
-            team.dayCount = day;
-            // enforce stats creation
-            getGlobalTeamStats(team, day);
-            getDayTeamStats(team, day);
+        // process day
+        if (!isCDF) {
+          data.forEach((match: any) => {
+            const teamA = getTeam(competition, match.EQA_no, match.EQA_nom);
+            const teamB = getTeam(competition, match.EQB_no, match.EQB_nom);
+            [teamA, teamB].forEach((team: Team) => {
+              dayCompetition.teams.push(team);
+              team.dayCount = day;
+              // enforce stats creation
+              getGlobalTeamStats(team, day);
+              getDayTeamStats(team, day);
+            });
           });
-        });
-      } else {
-        const poolCount = data.length / 3;
-        for (let i = 0; i < poolCount; i++) {
-          const m1 = data[3 * i];
-          const m2 = data[3 * i + 1];
-          // const m3 = data[3 * i + 2];
+        } else {
+          const poolCount = data.length / 3;
+          for (let i = 0; i < poolCount; i++) {
+            const m1 = data[3 * i];
+            const m2 = data[3 * i + 1];
+            // const m3 = data[3 * i + 2];
 
-          // find pool;
-          const teamA = getTeam(competition, m1.EQA_no, m1.EQA_nom);
-          const teamB = getTeam(competition, m1.EQB_no, m1.EQB_nom);
-          const teamC =
-            m2.EQA_no === m1.EQA_no || m2.EQA_no === m1.EQB_no
-              ? getTeam(competition, m2.EQB_no, m2.EQB_nom)
-              : getTeam(competition, m2.EQA_no, m2.EQA_nom);
-          const poolName = m1.Match.substring(1, 3);
-          let pool: Pool | undefined = dayCompetition.pools.get(poolName);
-          if (!pool) {
-            pool = {
-              name: poolName,
-              teams: [teamA, teamB, teamC],
-              matchs: [],
-            };
-            dayCompetition.pools.set(pool.name, pool);
+            // find pool;
+            const teamA = getTeam(competition, m1.EQA_no, m1.EQA_nom);
+            const teamB = getTeam(competition, m1.EQB_no, m1.EQB_nom);
+            const teamC =
+              m2.EQA_no === m1.EQA_no || m2.EQA_no === m1.EQB_no
+                ? getTeam(competition, m2.EQB_no, m2.EQB_nom)
+                : getTeam(competition, m2.EQA_no, m2.EQA_nom);
+            const poolName = m1.Match.substring(1, 3);
+            let pool: Pool | undefined = dayCompetition.pools.get(poolName);
+            if (!pool) {
+              pool = {
+                name: poolName,
+                teams: [teamA, teamB, teamC],
+                matchs: [],
+              };
+              dayCompetition.pools.set(pool.name, pool);
+            }
+            [teamA, teamB, teamC].forEach((team: Team) => {
+              dayCompetition.teams.push(team);
+              team.dayCount = day;
+              team.pools[day] = pool as Pool;
+              // enforce stats creation
+              getGlobalTeamStats(team, day);
+              getDayTeamStats(team, day);
+            });
           }
-          [teamA, teamB, teamC].forEach((team: Team) => {
-            dayCompetition.teams.push(team);
-            team.dayCount = day;
-            team.pools[day] = pool as Pool;
-            // enforce stats creation
-            getGlobalTeamStats(team, day);
-            getDayTeamStats(team, day);
-          });
         }
-      }
 
-      // process matchs
-      data.forEach((line: any) => {
-        const match = createMatch(competition, line);
-        if (isCDF) {
-          match.teamA.pools[day].matchs.push(match);
-        }
-        addCompetitionMatch(competition, match);
-      });
+        // process matchs
+        data.forEach((line: any) => {
+          const match = createMatch(competition, line);
+          if (isCDF) {
+            match.teamA.pools[day].matchs.push(match);
+          }
+          addCompetitionMatch(competition, match);
+        });
 
-      // compute ranking
-      const teams = Array.from(competition.teams.values());
-      teams.sort(rankingSorter(day, true));
-      teams.forEach((team, index) => {
-        team.ranking.globals[day] = index + 1;
-      });
-      const dteams = [...competition.days[day].teams];
-      dteams.sort(rankingSorter(day, false));
-      dteams.forEach((team, index) => {
-        team.ranking.days[day] = index + 1;
-      });
-      const running = [...competition.days[day].teams];
-      running.sort(rankingSorter(day, true));
-      running.forEach((team, index) => {
-        team.ranking.qualifieds[day] = index + 1;
-      });
-      competition.days[day].pools.forEach((pool: Pool) => {
-        const teams = [...pool.teams];
-        teams.sort(rankingSorter(day, false));
+        // compute ranking
+        const teams = Array.from(competition.teams.values());
+        teams.sort(rankingSorter(day, true));
         teams.forEach((team, index) => {
-          team.ranking.pools[day] = index + 1;
+          team.ranking.globals[day] = index + 1;
         });
+        const dteams = [...competition.days[day].teams];
+        dteams.sort(rankingSorter(day, false));
+        dteams.forEach((team, index) => {
+          team.ranking.days[day] = index + 1;
+        });
+        const running = [...competition.days[day].teams];
+        running.sort(rankingSorter(day, true));
+        running.forEach((team, index) => {
+          team.ranking.qualifieds[day] = index + 1;
+        });
+        competition.days[day].pools.forEach((pool: Pool) => {
+          const teams = [...pool.teams];
+          teams.sort(rankingSorter(day, false));
+          teams.forEach((team, index) => {
+            team.ranking.pools[day] = index + 1;
+          });
+        });
+        if (day > 1) {
+          const teams = [...competition.days[day].teams];
+          teams.sort(rankingSorter(day - 1, false));
+          teams.forEach((team: Team, index: number) => {
+            if (team.pools.length > 0 && team.pools[day].ranking === undefined) {
+              team.pools[day].ranking = index + 1;
+            }
+          });
+        }
       });
-      if (day > 1) {
-        const teams = [...competition.days[day].teams];
-        teams.sort(rankingSorter(day - 1, false));
-        teams.forEach((team: Team, index: number) => {
-          if (team.pools.length > 0 && team.pools[day].ranking === undefined) {
-            team.pools[day].ranking = index + 1;
-          }
-        });
-      }
     });
 };
